Make private-public key pair

Just for the convenience of it, make a key pair that may be used alongside sship. Please note that by default the private key will not be protected by a password.

sship_keygen(
  directory = normalizePath("~/.ssh"),
  type = "rsa",
  password = NULL
)

Arguments

directory

Character string with path to directory where the key pair will be written. Default is normalizePath("~/.ssh")

type

Character string defining the key type. Must be one of "rsa" (default) or "dsa". Key lengths are set to the default as defined in the openssl-package, currently 2048 and 1024, respectively

password

Character string with password to protect the private key. Default value is NULL in which case the private key will not be protected by a password

Value

Nothing will be returned from this function, but a message containing the directory where the keys were written is provided

Examples

sship_keygen(directory = tempdir())
#> sship: Key pair written to /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//RtmpWq5mFJ