Make calls to the github API

Provides a structured list of the specified resource from the the github API.

gh(path, proxy_url = NULL)

github_api(path, proxy_url = NULL)

rate_limit(proxy_url = NULL)

Arguments

path

Character string with path to the API resource.

proxy_url

Character string defining a network proxy in the form host:port. Default is NULL in which case the API call will not use a proxy.

Value

A list of class github_api containing the parsed content, API resource path and the response object. For rate_limit() the path is always "/rate_limit" and can hence be used to detect if the limit is exceeded (without being counted as a request itself). If the allowed API rate is exceeded gh() will return a message stating the fact and simple suggestions on how to remedy the problem.

Details

For most use cases only gh() will be relevant. The helper function github_api() do the actual lifting while rate_limit() handles API rate limits.

Examples

## Get all branches of a repository. If the api rate limit is exceeded this
## function will return NULL and an informative message
gh("repos/Rapporteket/sship/branches")
#> API rate limit is exceeded! You may have to wait until your limit is reset at 14:07:08 GMT or set env var GITHUB_PAT to your github personal access token to increase your limit.
#> NULL

## helper functions that will normally not be used
github_api("/rate_limit")
#> $content
#> $content$resources
#> $content$resources$core
#> $content$resources$core$limit
#> [1] 60
#> 
#> $content$resources$core$remaining
#> [1] 0
#> 
#> $content$resources$core$reset
#> [1] 1643378828
#> 
#> $content$resources$core$used
#> [1] 60
#> 
#> $content$resources$core$resource
#> [1] "core"
#> 
#> 
#> $content$resources$graphql
#> $content$resources$graphql$limit
#> [1] 0
#> 
#> $content$resources$graphql$remaining
#> [1] 0
#> 
#> $content$resources$graphql$reset
#> [1] 1643379477
#> 
#> $content$resources$graphql$used
#> [1] 0
#> 
#> $content$resources$graphql$resource
#> [1] "graphql"
#> 
#> 
#> $content$resources$integration_manifest
#> $content$resources$integration_manifest$limit
#> [1] 5000
#> 
#> $content$resources$integration_manifest$remaining
#> [1] 5000
#> 
#> $content$resources$integration_manifest$reset
#> [1] 1643379477
#> 
#> $content$resources$integration_manifest$used
#> [1] 0
#> 
#> $content$resources$integration_manifest$resource
#> [1] "integration_manifest"
#> 
#> 
#> $content$resources$search
#> $content$resources$search$limit
#> [1] 10
#> 
#> $content$resources$search$remaining
#> [1] 10
#> 
#> $content$resources$search$reset
#> [1] 1643375937
#> 
#> $content$resources$search$used
#> [1] 0
#> 
#> $content$resources$search$resource
#> [1] "search"
#> 
#> 
#> 
#> $content$rate
#> $content$rate$limit
#> [1] 60
#> 
#> $content$rate$remaining
#> [1] 0
#> 
#> $content$rate$reset
#> [1] 1643378828
#> 
#> $content$rate$used
#> [1] 60
#> 
#> $content$rate$resource
#> [1] "core"
#> 
#> 
#> 
#> $path
#> [1] "/rate_limit"
#> 
#> $response
#> Response [https://api.github.com/rate_limit]
#>   Date: 2022-01-28 13:17
#>   Status: 200
#>   Content-Type: application/json; charset=utf-8
#>   Size: 465 B
#> {"resources":{"core":{"limit":60,"remaining":0,"reset":1643378828,"used":60,"...
#> 
#> attr(,"class")
#> [1] "github_api"
rate_limit()
#> $content
#> $content$resources
#> $content$resources$core
#> $content$resources$core$limit
#> [1] 60
#> 
#> $content$resources$core$remaining
#> [1] 0
#> 
#> $content$resources$core$reset
#> [1] 1643378828
#> 
#> $content$resources$core$used
#> [1] 60
#> 
#> $content$resources$core$resource
#> [1] "core"
#> 
#> 
#> $content$resources$graphql
#> $content$resources$graphql$limit
#> [1] 0
#> 
#> $content$resources$graphql$remaining
#> [1] 0
#> 
#> $content$resources$graphql$reset
#> [1] 1643379477
#> 
#> $content$resources$graphql$used
#> [1] 0
#> 
#> $content$resources$graphql$resource
#> [1] "graphql"
#> 
#> 
#> $content$resources$integration_manifest
#> $content$resources$integration_manifest$limit
#> [1] 5000
#> 
#> $content$resources$integration_manifest$remaining
#> [1] 5000
#> 
#> $content$resources$integration_manifest$reset
#> [1] 1643379477
#> 
#> $content$resources$integration_manifest$used
#> [1] 0
#> 
#> $content$resources$integration_manifest$resource
#> [1] "integration_manifest"
#> 
#> 
#> $content$resources$search
#> $content$resources$search$limit
#> [1] 10
#> 
#> $content$resources$search$remaining
#> [1] 10
#> 
#> $content$resources$search$reset
#> [1] 1643375937
#> 
#> $content$resources$search$used
#> [1] 0
#> 
#> $content$resources$search$resource
#> [1] "search"
#> 
#> 
#> 
#> $content$rate
#> $content$rate$limit
#> [1] 60
#> 
#> $content$rate$remaining
#> [1] 0
#> 
#> $content$rate$reset
#> [1] 1643378828
#> 
#> $content$rate$used
#> [1] 60
#> 
#> $content$rate$resource
#> [1] "core"
#> 
#> 
#> 
#> $path
#> [1] "/rate_limit"
#> 
#> $response
#> Response [https://api.github.com/rate_limit]
#>   Date: 2022-01-28 13:17
#>   Status: 200
#>   Content-Type: application/json; charset=utf-8
#>   Size: 465 B
#> {"resources":{"core":{"limit":60,"remaining":0,"reset":1643378828,"used":60,"...
#> 
#> attr(,"class")
#> [1] "github_api"