Make calls to the github API

Provides a structured list of the specified resource from the the github API.

gh(path, proxy_url = NULL)

github_api(path, proxy_url = NULL)

rate_limit(proxy_url = NULL)

Arguments

path

Character string with path to the API resource.

proxy_url

Character string defining a network proxy in the form host:port. Default is NULL in which case the API call will not use a proxy.

Value

A list of class github_api containing the parsed content, API resource path and the response object. For rate_limit() the path is always "/rate_limit" and can hence be used to detect if the limit is exceeded (without being counted as a request itself). If the allowed API rate is exceeded gh() will return a message stating the fact and simple suggestions on how to remedy the problem.

Details

For most use cases only gh() will be relevant. The helper function github_api() do the actual lifting while rate_limit() handles API rate limits.

Examples

## Get all branches of a repository. If the api rate limit is exceeded this
## function will return NULL and an informative message
gh("repos/Rapporteket/sship/branches")
#> $content
#>                                           name
#> 1    dependabot/github_actions/actions/cache-3
#> 2 dependabot/github_actions/actions/checkout-3
#> 3    dependabot/github_actions/r-lib/actions-2
#> 4                                     gh-pages
#> 5                                         main
#>                                 commit.sha
#> 1 ef9f8d1f0435d45eceabdabe1dc2a60826c379dc
#> 2 089abed9bcea17fb7ec8aacbef77df08bf75ba85
#> 3 af1506f27879f0bdd552befa9c309c7e54fc7877
#> 4 9599ab4b7e7e3f86a44ea950dd9f04a8521b3364
#> 5 6a17a00ece0e01097dfdf74b946a3b8e80bf1dea
#>                                                                                        commit.url
#> 1 https://api.github.com/repos/Rapporteket/sship/commits/ef9f8d1f0435d45eceabdabe1dc2a60826c379dc
#> 2 https://api.github.com/repos/Rapporteket/sship/commits/089abed9bcea17fb7ec8aacbef77df08bf75ba85
#> 3 https://api.github.com/repos/Rapporteket/sship/commits/af1506f27879f0bdd552befa9c309c7e54fc7877
#> 4 https://api.github.com/repos/Rapporteket/sship/commits/9599ab4b7e7e3f86a44ea950dd9f04a8521b3364
#> 5 https://api.github.com/repos/Rapporteket/sship/commits/6a17a00ece0e01097dfdf74b946a3b8e80bf1dea
#>   protected
#> 1     FALSE
#> 2     FALSE
#> 3     FALSE
#> 4     FALSE
#> 5      TRUE
#> 
#> $path
#> [1] "repos/Rapporteket/sship/branches"
#> 
#> $response
#> Response [https://api.github.com/repos/Rapporteket/sship/branches]
#>   Date: 2022-05-13 07:41
#>   Status: 200
#>   Content-Type: application/json; charset=utf-8
#>   Size: 1.11 kB
#> 
#> 
#> attr(,"class")
#> [1] "github_api"

## helper functions that will normally not be used
github_api("/rate_limit")
#> $content
#> $content$resources
#> $content$resources$core
#> $content$resources$core$limit
#> [1] 60
#> 
#> $content$resources$core$remaining
#> [1] 43
#> 
#> $content$resources$core$reset
#> [1] 1652431056
#> 
#> $content$resources$core$used
#> [1] 17
#> 
#> $content$resources$core$resource
#> [1] "core"
#> 
#> 
#> $content$resources$graphql
#> $content$resources$graphql$limit
#> [1] 0
#> 
#> $content$resources$graphql$remaining
#> [1] 0
#> 
#> $content$resources$graphql$reset
#> [1] 1652431266
#> 
#> $content$resources$graphql$used
#> [1] 0
#> 
#> $content$resources$graphql$resource
#> [1] "graphql"
#> 
#> 
#> $content$resources$integration_manifest
#> $content$resources$integration_manifest$limit
#> [1] 5000
#> 
#> $content$resources$integration_manifest$remaining
#> [1] 5000
#> 
#> $content$resources$integration_manifest$reset
#> [1] 1652431266
#> 
#> $content$resources$integration_manifest$used
#> [1] 0
#> 
#> $content$resources$integration_manifest$resource
#> [1] "integration_manifest"
#> 
#> 
#> $content$resources$search
#> $content$resources$search$limit
#> [1] 10
#> 
#> $content$resources$search$remaining
#> [1] 10
#> 
#> $content$resources$search$reset
#> [1] 1652427726
#> 
#> $content$resources$search$used
#> [1] 0
#> 
#> $content$resources$search$resource
#> [1] "search"
#> 
#> 
#> 
#> $content$rate
#> $content$rate$limit
#> [1] 60
#> 
#> $content$rate$remaining
#> [1] 43
#> 
#> $content$rate$reset
#> [1] 1652431056
#> 
#> $content$rate$used
#> [1] 17
#> 
#> $content$rate$resource
#> [1] "core"
#> 
#> 
#> 
#> $path
#> [1] "/rate_limit"
#> 
#> $response
#> Response [https://api.github.com/rate_limit]
#>   Date: 2022-05-13 07:41
#>   Status: 200
#>   Content-Type: application/json; charset=utf-8
#>   Size: 467 B
#> {"resources":{"core":{"limit":60,"remaining":43,"reset":1652431056,"used":17,...
#> 
#> attr(,"class")
#> [1] "github_api"
rate_limit()
#> $content
#> $content$resources
#> $content$resources$core
#> $content$resources$core$limit
#> [1] 60
#> 
#> $content$resources$core$remaining
#> [1] 43
#> 
#> $content$resources$core$reset
#> [1] 1652431056
#> 
#> $content$resources$core$used
#> [1] 17
#> 
#> $content$resources$core$resource
#> [1] "core"
#> 
#> 
#> $content$resources$graphql
#> $content$resources$graphql$limit
#> [1] 0
#> 
#> $content$resources$graphql$remaining
#> [1] 0
#> 
#> $content$resources$graphql$reset
#> [1] 1652431266
#> 
#> $content$resources$graphql$used
#> [1] 0
#> 
#> $content$resources$graphql$resource
#> [1] "graphql"
#> 
#> 
#> $content$resources$integration_manifest
#> $content$resources$integration_manifest$limit
#> [1] 5000
#> 
#> $content$resources$integration_manifest$remaining
#> [1] 5000
#> 
#> $content$resources$integration_manifest$reset
#> [1] 1652431266
#> 
#> $content$resources$integration_manifest$used
#> [1] 0
#> 
#> $content$resources$integration_manifest$resource
#> [1] "integration_manifest"
#> 
#> 
#> $content$resources$search
#> $content$resources$search$limit
#> [1] 10
#> 
#> $content$resources$search$remaining
#> [1] 10
#> 
#> $content$resources$search$reset
#> [1] 1652427726
#> 
#> $content$resources$search$used
#> [1] 0
#> 
#> $content$resources$search$resource
#> [1] "search"
#> 
#> 
#> 
#> $content$rate
#> $content$rate$limit
#> [1] 60
#> 
#> $content$rate$remaining
#> [1] 43
#> 
#> $content$rate$reset
#> [1] 1652431056
#> 
#> $content$rate$used
#> [1] 17
#> 
#> $content$rate$resource
#> [1] "core"
#> 
#> 
#> 
#> $path
#> [1] "/rate_limit"
#> 
#> $response
#> Response [https://api.github.com/rate_limit]
#>   Date: 2022-05-13 07:41
#>   Status: 200
#>   Content-Type: application/json; charset=utf-8
#>   Size: 467 B
#> {"resources":{"core":{"limit":60,"remaining":43,"reset":1652431056,"used":17,...
#> 
#> attr(,"class")
#> [1] "github_api"